+++ json_serializable/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 3.4.2-dev
+
+- `JsonKey.unknownEnumValue`
+  - Added support for `unknownEnumValue` on `List` of enums
+
 ## 3.4.1
 
 - Support properties where the getter is defined in a class with a corresponding